+++ src/azure-cli/azure/cli/command_modules/cosmosdb/_params.py
@@ -102,7 +102,7 @@ def load_arguments(self, _):
     database_name_type = CLIArgumentType(options_list=['--database-name', '-d'], help='Database name.')
     container_name_type = CLIArgumentType(options_list=['--container-name', '-c'], help='Container name.')
 
-    with self.argument_context('cosmosdb private-endpoint') as c:
+    with self.argument_context('cosmosdb private-endpoint-connection') as c:
         c.argument('approval_description', help='Comments for the approval.')
         c.argument('private_endpoint_connection_name', options_list=['--connection-name', '-n'], required=False,
                    help='The name of the private endpoint connection associated with Azure Cosmos DB. '
@@ -114,7 +114,7 @@ def load_arguments(self, _):
         c.argument('rejection_description', help='Comments for the rejection.')
 
     for item in ['approve', 'reject', 'delete', 'show']:
-        with self.argument_context('cosmosdb private-endpoint {}'.format(item)) as c:
+        with self.argument_context('cosmosdb private-endpoint-connection {}'.format(item)) as c:
             c.extra('connection_id', required=False,
                     help='The ID of the private endpoint connection associated with Azure Cosmos DB. '
                          'If specified --account-name and --connection-name/-n, this should be omitted.')